+++ MatrixColumnCount.java
@@ -1,5 +1,5 @@
 public class MatrixColumnCount {
-    public static int matrixColumnCount(int[][] matrix_i) {
+    public static int matrixColumnCount(float[][] matrix_i) {
         return matrix_i[0].length;
 
     }
