+++ MatrixMultiply.java
@@ -1,7 +1,7 @@
 public class MatrixMultiply {
-    public static int[][] matrixMultiply(int[][] matrix_x, int[][] matrix_y) {
-        int[][] matrix_a = matrix_x;
-        int[][] matrix_b = matrix_y;
+    public static float[][] matrixMultiply(float[][] matrix_x, float[][] matrix_y) {
+        float[][] matrix_a = matrix_x;
+        float[][] matrix_b = matrix_y;
         if (!(MatrixColumnCount.matrixColumnCount(matrix_a) == MatrixRowCount.matrixRowCount(matrix_b))) {
             if (MatrixColumnCount.matrixColumnCount(matrix_b) == MatrixRowCount.matrixRowCount(matrix_a)) {
                 matrix_a = matrix_y;
@@ -11,7 +11,7 @@ public class MatrixMultiply {
                 return matrix_x;
             }
         }
-        int[][] matrix_p = new int[MatrixRowCount.matrixRowCount(matrix_a)][MatrixColumnCount
+        float[][] matrix_p = new float[MatrixRowCount.matrixRowCount(matrix_a)][MatrixColumnCount
                 .matrixColumnCount(matrix_b)];
         for (int i = 0; i < matrix_p.length; i++) {
             System.out.println("multiplying...");
@@ -20,12 +20,12 @@ public class MatrixMultiply {
         return matrix_p;
     }
 
-    public static int[] multiplyTuples(int[][] matrix_a, int[][] matrix_b, int row) {
-        int val = 0;
-        int[] tuple = new int[MatrixColumnCount.matrixColumnCount(matrix_b)];
+    public static float[] multiplyTuples(float[][] matrix_a, float[][] matrix_b, int row) {
+        float val = 0;
+        float[] tuple = new float[MatrixColumnCount.matrixColumnCount(matrix_b)];
         for (int i = 0; i < matrix_b[0].length; i++) {
             for (int j = 0; j < matrix_a[row - 1].length; j++) {
-                int temp = matrix_a[row - 1][j] * matrix_b[j][i];
+                float temp = matrix_a[row - 1][j] * matrix_b[j][i];
                 val += temp;
             }
             tuple[i] = val;
