+++ MatrixRowCount.java
@@ -1,6 +1,6 @@
 public class MatrixRowCount {
 
-    public static int matrixRowCount(int[][] matrix_i) {
+    public static int matrixRowCount(float[][] matrix_i) {
         return matrix_i.length;
 
     }
